# exam_api/api/docs/swagger.yaml
definitions:
  custumer.Addresss:
    properties:
      home_adress:
        type: string
      street:
        type: string
    type: object
  custumer.CustumAddress:
    properties:
      home_adress:
        type: string
      id:
        type: integer
      street:
        type: string
    type: object
  custumer.CustumerForCreate:
    properties:
      addresses:
        items:
          $ref: '#/definitions/custumer.Addresss'
        type: array
      bio:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phoneNumber:
        type: string
    type: object
  custumer.CustumerInfo:
    properties:
      Bio:
        type: string
      addresses:
        items:
          $ref: '#/definitions/custumer.CustumAddress'
        type: array
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      phoneNumber:
        type: string
    type: object
  custumer.Empty:
    type: object
  post.EmptyPost:
    type: object
  post.Media:
    properties:
      link:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  post.PostForCreate:
    properties:
      description_post:
        type: string
      medias:
        items:
          $ref: '#/definitions/post.Media'
        type: array
      poster_id:
        type: integer
    type: object
  post.PostInfo:
    properties:
      description_post:
        type: string
      id:
        type: integer
      medias:
        items:
          $ref: '#/definitions/post.Media'
        type: array
      poster_id:
        type: integer
    type: object
  reating.EmptyReating:
    type: object
  reating.ReatingForCreate:
    properties:
      custumer_id:
        type: integer
      description:
        type: string
      post_id:
        type: integer
      reating:
        type: integer
    type: object
  reating.ReatingInfo:
    properties:
      custumer_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      post_id:
        type: integer
      reating:
        type: integer
    type: object
  reating.Reatings:
    properties:
      reatins:
        items:
          $ref: '#/definitions/reating.ReatingInfo'
        type: array
    type: object
info:
  contact: {}
paths:
  /v1/custumer/create:
    post:
      consumes:
      - application/json
      description: this api create store
      parameters:
      - description: Custumer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/custumer.CustumerForCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/custumer.CustumerInfo'
      summary: create store api
      tags:
      - Custumer
  /v1/custumer/delete/{id}:
    delete:
      consumes:
      - application/json
      description: this api delet custumer with posts by id
      parameters:
      - description: Custumer id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/custumer.Empty'
      summary: delete Custumer api
      tags:
      - Custumer
  /v1/custumer/get/{id}:
    get:
      consumes:
      - application/json
      description: this api get custumer with posts by id
      parameters:
      - description: Custumer id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/custumer.CustumerInfo'
      summary: get Custumer api
      tags:
      - Custumer
  /v1/custumer/update:
    post:
      consumes:
      - application/json
      description: this api update custumer
      parameters:
      - description: Custumer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/custumer.CustumerInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/custumer.CustumerInfo'
      summary: update custumer api
      tags:
      - Custumer
  /v1/post/allInfo/{id}:
    get:
      consumes:
      - application/json
      description: this api get Post by id
      parameters:
      - description: Poster id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: object
      summary: get Post api
      tags:
      - Post
  /v1/post/create:
    post:
      consumes:
      - application/json
      description: this api create post
      parameters:
      - description: Custumer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/post.PostForCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PostInfo'
      summary: create post api
      tags:
      - Post
  /v1/post/delet/{id}:
    delete:
      consumes:
      - application/json
      description: this api posts by id
      parameters:
      - description: Post id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.EmptyPost'
      summary: delete Post api
      tags:
      - Post
  /v1/post/get/reatings/{id}:
    get:
      consumes:
      - application/json
      description: this api get posts reating by id
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reating.Reatings'
      summary: Get  posts reating api
      tags:
      - Post
  /v1/post/update:
    post:
      consumes:
      - application/json
      description: this api update Post
      parameters:
      - description: Post
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/post.PostInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/post.PostInfo'
      summary: update post api
      tags:
      - Post
  /v1/reating/create:
    post:
      consumes:
      - application/json
      description: this api create reating
      parameters:
      - description: Custumer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reating.ReatingForCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reating.ReatingInfo'
      summary: create reating api
      tags:
      - Reating
  /v1/reating/delete/{id}:
    delete:
      consumes:
      - application/json
      description: this api delet reating by id
      parameters:
      - description: reating id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reating.EmptyReating'
      summary: delete Reating api
      tags:
      - Reating
  /v1/reating/get/{id}:
    get:
      consumes:
      - application/json
      description: this api get reating by id
      parameters:
      - description: reating id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reating.ReatingInfo'
      summary: Get reating api
      tags:
      - Reating
  /v1/reating/update:
    post:
      consumes:
      - application/json
      description: this api update reating
      parameters:
      - description: reating
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/reating.ReatingInfo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/reating.ReatingInfo'
      summary: update reating api
      tags:
      - Reating
swagger: "2.0"